Remove Control Panel Links
Remove control panel links to avoid customer confusion. This article applies to admins who should have some experience with in the cp. Providing links for SA3 and Core accounts can lead to confusion because there is no clear way to explain to the customer which billing system their account may exist in.